Implemented observer pattern to reload red-glass.js whenever it is cleared from the browser's memory.
Whenever a new page is loaded, red-glass.js is removed from the
browser's memory- so no additional events can be recorded from that
point forward.  This commit is the first pass at adding an observer
pattern to WebDriver.  The intended goal is for RedGlass to detect when
its javascript app is no longer in memory, and subsequently reload it.
I have only added notifiers to the Navigation#to and Element#click
methods so far- more to follow.