Return parent element if target element is blank.
A mutation event (DOM node added/removed) will now return the inner
HTML of the parent element if the HTML of the target element is blank.
This will provide a bit more context when looking at the detailed view
of a mutation event record.  Instead of seeing a blank string you will
now see the parent element.