Added simple event sequence recording.
During an automation session, we want to keep track of the sequence of
events which transitioned a page into a particular state. This basic
sequence recorder only indicates an event type and an element tag name.

This functionality is intended to complement the event recording
accomplished by the JavaScript monitor in the browser.